QuieroAprender/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.3
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3

django-ckeditor~=6.7.1
requests~=2.32.3
python-dotenv~=1.0.1